+++ src/components/global/logo/logo.tsx
@@ -1,5 +1,5 @@
 import {Link} from 'react-router-dom';
-import {AppRoute} from '../../types/routes';
+import {AppRoute} from '../../../types/routes';
 
 export function Logo() {
   return (
