+++ src/components/global/private-route/private-route.tsx
@@ -1,5 +1,5 @@
 import {Navigate} from 'react-router-dom';
-import {AppRoute, AppRouteProps, AuthorizationStatus} from '../../types/routes';
+import {AppRoute, AppRouteProps, AuthorizationStatus} from '../../../types/routes';
 
 export function PrivateRoute({authorizationStatus, children}: AppRouteProps) {
   return (
