+++ src/pages/page-favorites-empty/page-favorites-empty.tsx
@@ -1,36 +1,10 @@
-import {Logo} from '../../components/logo/logo';
 import {Helmet} from 'react-helmet-async';
 
 function PageFavoritesEmpty() {
   return (
-    <div className="page page--favorites-empty">
+    <>
+      {/* На этой странице в .page добавляются классы "page--favorites-empty" */}
       <Helmet><title>6 cities. Favorites</title></Helmet>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Logo/>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <span className="header__favorite-count">0</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="#">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
 
       <main className="page__main page__main--favorites page__main--favorites-empty">
         <div className="page__favorites-container container">
@@ -48,7 +22,7 @@ function PageFavoritesEmpty() {
           <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33"/>
         </a>
       </footer>
-    </div>
+    </>
   );
 }
