+++ src/pages/page-favorites/page-favorites.tsx
@@ -1,36 +1,9 @@
 import {Helmet} from 'react-helmet-async';
-import {Logo} from '../../components/logo/logo';
 
 function PageFavorites() {
   return (
-    <div className="page">
+    <>
       <Helmet><title>6 cities. Favorites</title></Helmet>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Logo/>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <span className="header__favorite-count">3</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="#">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
 
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
@@ -166,7 +139,7 @@ function PageFavorites() {
           <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33"/>
         </a>
       </footer>
-    </div>
+    </>
   );
 }
