+++ src/pages/page-main-empty/page-main-empty.tsx
@@ -1,36 +1,11 @@
-import {LogoActive} from '../../components/logo/logo';
+import {LogoActive} from '../../components/global/logo/logo';
 import {Helmet} from 'react-helmet-async';
 
 function PageMainEmpty() {
   return (
-    <div className="page page--gray page--main">
+    <>
+      {/* на этой стринице у .page должны быть классы "page page--gray page--main" */}
       <Helmet><title>6 cities</title></Helmet>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <LogoActive/>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <span className="header__favorite-count">3</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="#">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
 
       <main className="page__main page__main--index page__main--index-empty">
         <h1 className="visually-hidden">Cities</h1>
@@ -82,7 +57,7 @@ function PageMainEmpty() {
           </div>
         </div>
       </main>
-    </div>
+    </>
   );
 }
