+++ src/pages/page-main/page-main.tsx
@@ -1,39 +1,13 @@
 import {PlaceCard} from '../../components/place-card/place-card';
 import {PageMainProps} from '../../types/common-types';
 import {OFFERS} from '../../mock/offers';
-import {LogoActive} from '../../components/logo/logo';
 import {Helmet} from 'react-helmet-async';
 
 function PageMain({placesCount}: PageMainProps) {
   return (
-    <div className="page page--gray page--main">
+    <>
+      {/* На этой странице при открытии добавляются классы "page--gray page--main" */}
       <Helmet><title>6 cities</title></Helmet>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <LogoActive/>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <span className="header__favorite-count">3</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="#">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
 
       <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
@@ -103,7 +77,7 @@ function PageMain({placesCount}: PageMainProps) {
           </div>
         </div>
       </main>
-    </div>
+    </>
   );
 }
