+++ src/pages/page-offer-not-logged/page-offer-not-logged.tsx
@@ -1,30 +1,9 @@
-import {Logo} from '../../components/logo/logo';
 import {Helmet} from 'react-helmet-async';
 
 function PageOfferNotLogged() {
   return (
-    <div className="page">
+    <>
       <Helmet><title>6 cities. Offer</title></Helmet>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Logo/>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__login">Sign in</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
 
       <main className="page__main page__main--offer">
         <section className="offer">
@@ -282,7 +261,7 @@ function PageOfferNotLogged() {
           </section>
         </div>
       </main>
-    </div>
+    </>
   );
 }
