+++ src/pages/page-offer/page-offer.tsx
@@ -1,36 +1,9 @@
-import {Logo} from '../../components/logo/logo';
 import {Helmet} from 'react-helmet-async';
 
 function PageOffer() {
   return (
-    <div className="page">
+    <>
       <Helmet><title>6 cities. Offer</title></Helmet>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Logo/>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <span className="header__favorite-count">3</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="#">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
 
       <main className="page__main page__main--offer">
         <section className="offer">
@@ -334,7 +307,7 @@ function PageOffer() {
           </section>
         </div>
       </main>
-    </div>
+    </>
   );
 }
